+++ src/data/api-categories.mjs
@@ -1,15 +1,13 @@
 // mapping of api categories coming in from libraries to the associated categories in the docs
-export const packageCategories = {
-  'amplify-js': {
-    ROOT_PACKAGE: 'aws-amplify',
-    API_CATEGORIES: {
-      auth: 'auth',
-      storage: 'storage'
-    },
-    API_SUB_CATEGORIES: {
-      analytics: 'analytics',
-      'rest-api': 'api',
-      'in-app-messaging': 'in-app-messaging'
-    }
-  }
+export const API_CATEGORIES = {
+  auth: 'auth',
+  storage: 'storage'
 };
+
+export const API_SUB_CATEGORIES = {
+  analytics: 'analytics',
+  'rest-api': 'api',
+  'in-app-messaging': 'in-app-messaging'
+};
+
+export const ROOT_PACKAGE = 'amplify-js';
