+++ src/utils/getApiStaticPath.ts
@@ -1,7 +1,5 @@
-import { packageCategories } from '@/data/api-categories';
+import { API_CATEGORIES, API_SUB_CATEGORIES } from '@/data/api-categories.mjs';
 import { JS_PLATFORMS } from '@/data/platforms';
-
-const { API_CATEGORIES, API_SUB_CATEGORIES } = packageCategories['amplify-js'];
 
 export const getApiStaticPath = (sub) => {
   const paths: any = [];
